DigiCart/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-jazzmin==2.5.0
django-on-heroku==1.1.2
django-phonenumber-field==6.3.0
gunicorn==20.1.0
idna==3.3
phonenumberslite==8.12.51
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.2.0
